Enhance the resolver plugin with option to eliminate afterEvaluate usage.
The afterEvaluate usage of the resolver plugin imposes undesirable
side effects to the resolution strategy. Making the configuration
as a two step plugin apply operation is a functionally equivalent
solution that is also compatible with existing scripting.

Applying the plugin as it always was done will still use the
afterEvaluate operation. If the new optional plugin called
version-resolver-config is applied first the extensions are
available to inject configuration parameters before the plugin
is applied itself. When this approach is used the plugin will
use an immediate strategy instead of the afterEvaluate approach.